+++ benches/bench_main.rs
@@ -12,4 +12,5 @@ criterion_main! {
     benchmarks::special_characters::benches,
     benchmarks::measurement_overhead::benches,
     benchmarks::custom_measurement::benches,
+    benchmarks::sampling_mode::benches,
 }