+++ benches/benchmarks/mod.rs
@@ -5,5 +5,6 @@ pub mod iter_with_large_drop;
 pub mod iter_with_large_setup;
 pub mod iter_with_setup;
 pub mod measurement_overhead;
+pub mod sampling_mode;
 pub mod special_characters;
 pub mod with_inputs;