+++ src/connection.rs
@@ -359,3 +359,11 @@ pub enum SamplingMethod {
     Linear,
     Flat,
 }
+impl From<crate::ActualSamplingMode> for SamplingMethod {
+    fn from(other: crate::ActualSamplingMode) -> Self {
+        match other {
+            crate::ActualSamplingMode::Flat => SamplingMethod::Flat,
+            crate::ActualSamplingMode::Linear => SamplingMethod::Linear,
+        }
+    }
+}