+++ src/estimate.rs
@@ -65,7 +65,7 @@ pub fn build_estimates(
         mean: to_estimate(points.mean, &distributions.mean),
         median: to_estimate(points.median, &distributions.median),
         median_abs_dev: to_estimate(points.median_abs_dev, &distributions.median_abs_dev),
-        slope: to_estimate(points.slope, &distributions.slope),
+        slope: None,
         std_dev: to_estimate(points.std_dev, &distributions.std_dev),
     }
 }
@@ -99,7 +99,6 @@ pub struct PointEstimates {
     pub mean: f64,
     pub median: f64,
     pub median_abs_dev: f64,
-    pub slope: f64,
     pub std_dev: f64,
 }
 
@@ -108,21 +107,21 @@ pub struct Estimates {
     pub mean: Estimate,
     pub median: Estimate,
     pub median_abs_dev: Estimate,
-    pub slope: Estimate,
+    pub slope: Option<Estimate>,
     pub std_dev: Estimate,
 }
 impl Estimates {
     pub fn typical(&self) -> &Estimate {
-        &self.slope
+        self.slope.as_ref().unwrap_or(&self.mean)
     }
-    pub fn get(&self, stat: Statistic) -> &Estimate {
+    pub fn get(&self, stat: Statistic) -> Option<&Estimate> {
         match stat {
-            Statistic::Mean => &self.mean,
-            Statistic::Median => &self.median,
-            Statistic::MedianAbsDev => &self.median_abs_dev,
-            Statistic::Slope => &self.slope,
-            Statistic::StdDev => &self.std_dev,
-            Statistic::Typical => self.typical(),
+            Statistic::Mean => Some(&self.mean),
+            Statistic::Median => Some(&self.median),
+            Statistic::MedianAbsDev => Some(&self.median_abs_dev),
+            Statistic::Slope => self.slope.as_ref(),
+            Statistic::StdDev => Some(&self.std_dev),
+            Statistic::Typical => Some(self.typical()),
         }
     }
 }
@@ -131,21 +130,21 @@ pub struct Distributions {
     pub mean: Distribution<f64>,
     pub median: Distribution<f64>,
     pub median_abs_dev: Distribution<f64>,
-    pub slope: Distribution<f64>,
+    pub slope: Option<Distribution<f64>>,
     pub std_dev: Distribution<f64>,
 }
 impl Distributions {
     pub fn typical(&self) -> &Distribution<f64> {
-        &self.slope
+        self.slope.as_ref().unwrap_or(&self.mean)
     }
-    pub fn get(&self, stat: Statistic) -> &Distribution<f64> {
+    pub fn get(&self, stat: Statistic) -> Option<&Distribution<f64>> {
         match stat {
-            Statistic::Mean => &self.mean,
-            Statistic::Median => &self.median,
-            Statistic::MedianAbsDev => &self.median_abs_dev,
-            Statistic::Slope => &self.slope,
-            Statistic::StdDev => &self.std_dev,
-            Statistic::Typical => self.typical(),
+            Statistic::Mean => Some(&self.mean),
+            Statistic::Median => Some(&self.median),
+            Statistic::MedianAbsDev => Some(&self.median_abs_dev),
+            Statistic::Slope => self.slope.as_ref(),
+            Statistic::StdDev => Some(&self.std_dev),
+            Statistic::Typical => Some(self.typical()),
         }
     }
 }