+++ src/html/mod.rs
@@ -6,6 +6,7 @@ use crate::format;
 use crate::fs;
 use crate::measurement::ValueFormatter;
 use crate::plot::{PlotContext, PlotData, Plotter};
+use crate::SavedSample;
 use criterion_plot::Size;
 use serde::Serialize;
 use std::cell::RefCell;
@@ -36,7 +37,7 @@ struct Context {
     thumbnail_width: usize,
     thumbnail_height: usize,
 
-    slope: ConfidenceInterval,
+    slope: Option<ConfidenceInterval>,
     r2: ConfidenceInterval,
     mean: ConfidenceInterval,
     std_dev: ConfidenceInterval,
@@ -53,12 +54,23 @@ struct Context {
 struct IndividualBenchmark {
     name: String,
     path: String,
+    regression_exists: bool,
 }
 impl IndividualBenchmark {
-    fn from_id(path_prefix: &str, id: &BenchmarkId) -> IndividualBenchmark {
+    fn from_id(
+        output_directory: &Path,
+        path_prefix: &str,
+        id: &BenchmarkId,
+    ) -> IndividualBenchmark {
+        let mut regression_path = PathBuf::from(output_directory);
+        regression_path.push(id.as_directory_name());
+        regression_path.push("report");
+        regression_path.push("regression.svg");
+
         IndividualBenchmark {
             name: id.as_title().to_owned(),
             path: format!("{}/{}", path_prefix, id.as_directory_name()),
+            regression_exists: regression_path.is_file(),
         }
     }
 }
@@ -295,7 +307,7 @@ impl Report for Html {
             fs::mkdirp(&report_dir)
         });
 
-        let slope_estimate = &measurements.absolute_estimates.slope;
+        let typical_estimate = &measurements.absolute_estimates.typical();
 
         let time_interval = |est: &Estimate| -> ConfidenceInterval {
             ConfidenceInterval {
@@ -312,25 +324,43 @@ impl Report for Html {
             self.generate_plots(id, report_context, formatter, measurements)
         }
 
+        let mut additional_plots = vec![
+            Plot::new("Typical", "typical.svg"),
+            Plot::new("Mean", "mean.svg"),
+            Plot::new("Std. Dev.", "SD.svg"),
+            Plot::new("Median", "median.svg"),
+            Plot::new("MAD", "MAD.svg"),
+        ];
+        if measurements.absolute_estimates.slope.is_some() {
+            additional_plots.push(Plot::new("Slope", "slope.svg"));
+        }
+
         let throughput = measurements
             .throughput
             .as_ref()
             .map(|thr| ConfidenceInterval {
                 lower: formatter
-                    .format_throughput(thr, slope_estimate.confidence_interval.upper_bound),
+                    .format_throughput(thr, typical_estimate.confidence_interval.upper_bound),
                 upper: formatter
-                    .format_throughput(thr, slope_estimate.confidence_interval.lower_bound),
-                point: formatter.format_throughput(thr, slope_estimate.point_estimate),
+                    .format_throughput(thr, typical_estimate.confidence_interval.lower_bound),
+                point: formatter.format_throughput(thr, typical_estimate.point_estimate),
             });
 
         let context = Context {
             title: id.as_title().to_owned(),
-            confidence: format!("{:.2}", slope_estimate.confidence_interval.confidence_level),
+            confidence: format!(
+                "{:.2}",
+                typical_estimate.confidence_interval.confidence_level
+            ),
 
             thumbnail_width: THUMBNAIL_SIZE.unwrap().0,
             thumbnail_height: THUMBNAIL_SIZE.unwrap().1,
 
-            slope: time_interval(slope_estimate),
+            slope: measurements
+                .absolute_estimates
+                .slope
+                .as_ref()
+                .map(time_interval),
             mean: time_interval(&measurements.absolute_estimates.mean),
             median: time_interval(&measurements.absolute_estimates.median),
             mad: time_interval(&measurements.absolute_estimates.median_abs_dev),
@@ -340,25 +370,19 @@ impl Report for Html {
             r2: ConfidenceInterval {
                 lower: format!(
                     "{:0.7}",
-                    Slope(slope_estimate.confidence_interval.lower_bound).r_squared(&data)
+                    Slope(typical_estimate.confidence_interval.lower_bound).r_squared(&data)
                 ),
                 upper: format!(
                     "{:0.7}",
-                    Slope(slope_estimate.confidence_interval.upper_bound).r_squared(&data)
+                    Slope(typical_estimate.confidence_interval.upper_bound).r_squared(&data)
                 ),
                 point: format!(
                     "{:0.7}",
-                    Slope(slope_estimate.point_estimate).r_squared(&data)
+                    Slope(typical_estimate.point_estimate).r_squared(&data)
                 ),
             },
 
-            additional_plots: vec![
-                Plot::new("Slope", "slope.svg"),
-                Plot::new("Mean", "mean.svg"),
-                Plot::new("Std. Dev.", "SD.svg"),
-                Plot::new("Median", "median.svg"),
-                Plot::new("MAD", "MAD.svg"),
-            ],
+            additional_plots,
 
             comparison: self.comparison(measurements),
         };
@@ -389,6 +413,9 @@ impl Report for Html {
                 fs::is_dir(&id_dir)
             })
             .collect::<Vec<_>>();
+        if all_ids.is_empty() {
+            return;
+        }
 
         let group_id = all_ids[0].group_id.clone();
 
@@ -627,10 +654,19 @@ impl Html {
 
         self.plotter.borrow_mut().pdf(plot_ctx, plot_data);
         self.plotter.borrow_mut().pdf(plot_ctx_small, plot_data);
-        self.plotter.borrow_mut().regression(plot_ctx, plot_data);
-        self.plotter
-            .borrow_mut()
-            .regression(plot_ctx_small, plot_data);
+        if measurements.absolute_estimates.slope.is_some() {
+            self.plotter.borrow_mut().regression(plot_ctx, plot_data);
+            self.plotter
+                .borrow_mut()
+                .regression(plot_ctx_small, plot_data);
+        } else {
+            self.plotter
+                .borrow_mut()
+                .iteration_times(plot_ctx, plot_data);
+            self.plotter
+                .borrow_mut()
+                .iteration_times(plot_ctx_small, plot_data);
+        }
 
         self.plotter
             .borrow_mut()
@@ -657,10 +693,21 @@ impl Html {
 
             self.plotter.borrow_mut().pdf(plot_ctx, comp_data);
             self.plotter.borrow_mut().pdf(plot_ctx_small, comp_data);
-            self.plotter.borrow_mut().regression(plot_ctx, comp_data);
-            self.plotter
-                .borrow_mut()
-                .regression(plot_ctx_small, comp_data);
+            if measurements.absolute_estimates.slope.is_some()
+                && comp.base_estimates.slope.is_some()
+            {
+                self.plotter.borrow_mut().regression(plot_ctx, comp_data);
+                self.plotter
+                    .borrow_mut()
+                    .regression(plot_ctx_small, comp_data);
+            } else {
+                self.plotter
+                    .borrow_mut()
+                    .iteration_times(plot_ctx, comp_data);
+                self.plotter
+                    .borrow_mut()
+                    .iteration_times(plot_ctx_small, comp_data);
+            }
             self.plotter.borrow_mut().t_test(plot_ctx, comp_data);
             self.plotter
                 .borrow_mut()
@@ -680,7 +727,7 @@ impl Html {
             .filter_map(|id| {
                 let entry = output_directory.join(id.as_directory_name()).join("new");
 
-                let (iters, times): (Vec<f64>, Vec<f64>) =
+                let SavedSample { iters, times, .. } =
                     try_else_return!(fs::load(&entry.join("sample.json")), || None);
                 let avg_times = iters
                     .into_iter()
@@ -738,7 +785,9 @@ impl Html {
         let path_prefix = if full_summary { "../.." } else { "../../.." };
         let benchmarks = data
             .iter()
-            .map(|&&(ref id, _)| IndividualBenchmark::from_id(path_prefix, id))
+            .map(|&&(ref id, _)| {
+                IndividualBenchmark::from_id(&report_context.output_directory, path_prefix, id)
+            })
             .collect();
 
         let context = SummaryContext {