+++ src/plot/mod.rs
@@ -9,11 +9,12 @@ use crate::measurement::ValueFormatter;
 use crate::report::{BenchmarkId, ComparisonData, MeasurementData, ReportContext, ValueType};
 use std::path::PathBuf;
 
-const REPORT_STATS: [Statistic; 5] = [
+const REPORT_STATS: [Statistic; 6] = [
+    Statistic::Typical,
+    Statistic::Slope,
     Statistic::Mean,
     Statistic::Median,
     Statistic::MedianAbsDev,
-    Statistic::Slope,
     Statistic::MedianAbsDev,
 ];
 const CHANGE_STATS: [Statistic; 2] = [Statistic::Mean, Statistic::Median];
@@ -73,6 +74,8 @@ pub(crate) trait Plotter {
     fn pdf(&mut self, ctx: PlotContext<'_>, data: PlotData<'_>);
 
     fn regression(&mut self, ctx: PlotContext<'_>, data: PlotData<'_>);
+
+    fn iteration_times(&mut self, ctx: PlotContext<'_>, data: PlotData<'_>);
 
     fn abs_distributions(&mut self, ctx: PlotContext<'_>, data: PlotData<'_>);