+++ src/plot/plotters_backend/distributions.rs
@@ -126,17 +126,27 @@ pub(crate) fn abs_distributions(
     measurements: &MeasurementData<'_>,
     size: Option<(u32, u32)>,
 ) {
-    crate::plot::REPORT_STATS.iter().for_each(|&statistic| {
-        abs_distribution(
-            id,
-            context,
-            formatter,
-            statistic,
-            measurements.distributions.get(statistic),
-            measurements.absolute_estimates.get(statistic),
-            size,
-        )
-    })
+    crate::plot::REPORT_STATS
+        .iter()
+        .filter_map(|stat| {
+            measurements.distributions.get(*stat).and_then(|dist| {
+                measurements
+                    .absolute_estimates
+                    .get(*stat)
+                    .map(|est| (*stat, dist, est))
+            })
+        })
+        .for_each(|(statistic, distribution, estimate)| {
+            abs_distribution(
+                id,
+                context,
+                formatter,
+                statistic,
+                distribution,
+                estimate,
+                size,
+            )
+        })
 }
 
 fn rel_distribution(