+++ src/plot/plotters_backend/regression.rs
@@ -13,8 +13,8 @@ pub(crate) fn regression_figure(
     measurements: &MeasurementData<'_>,
     size: Option<(u32, u32)>,
 ) {
-    let slope_estimate = &measurements.absolute_estimates.slope;
-    let slope_dist = &measurements.distributions.slope;
+    let slope_estimate = measurements.absolute_estimates.slope.as_ref().unwrap();
+    let slope_dist = measurements.distributions.slope.as_ref().unwrap();
     let (lb, ub) =
         slope_dist.confidence_interval(slope_estimate.confidence_interval.confidence_level);
 
@@ -141,7 +141,7 @@ pub(crate) fn regression_comparison_figure(
             },
         point_estimate: base_point,
         ..
-    } = comparison.base_estimates.slope;
+    } = comparison.base_estimates.slope.as_ref().unwrap();
 
     let Estimate {
         confidence_interval:
@@ -152,7 +152,7 @@ pub(crate) fn regression_comparison_figure(
             },
         point_estimate: point,
         ..
-    } = comparison.base_estimates.slope;
+    } = measurements.absolute_estimates.slope.as_ref().unwrap();
 
     let mut points = [
         base_lb * max_iters,